--- src/main/java/com/trainingdeveloperpro/k001/Main.java
@@ -1,4 +1,0 @@
-package com.trainingdeveloperpro.k001;
-
-public class Main {
-}
