--- src/main/java/com/trainingdeveloperpro/k001/dokhuongduy/Main.java
@@ -1,4 +1,0 @@
-package com.trainingdeveloperpro.k001.dokhuongduy;
-
-public class Main {
-}
