--- src/main/java/com/trainingdeveloperpro/k001/hoangtrieuthang/Main.java
@@ -1,4 +1,0 @@
-package com.trainingdeveloperpro.k001.hoangtrieuthang;
-
-public class Main {
-}
