--- src/main/java/com/trainingdeveloperpro/k001/nguyentrungkhanh/Main.java
@@ -1,4 +1,0 @@
-package com.trainingdeveloperpro.k001.nguyentrungkhanh;
-
-public class Main {
-}
