--- src/main/java/com/trainingdeveloperpro/k001/phamminhquoc/Main.java
@@ -1,4 +1,0 @@
-package com.trainingdeveloperpro.k001.phamminhquoc;
-
-public class Main {
-}
