--- src/main/java/com/trainingdeveloperpro/k001/phamtuananh/Main.java
@@ -1,4 +1,0 @@
-package com.trainingdeveloperpro.k001.phamtuananh;
-
-public class Main {
-}
