--- src/main/resources/beans.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<beans xmlns="http://www.springframework.org/schema/beans"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
-
-    <bean id = "helloWorld1" class="com.trainingdeveloperpro.k001.danghuucanh.HelloWorld">
-        <property name="message" value="inject by setter"/>
-    </bean>
-
-    <bean id="helloWorld2" class="com.trainingdeveloperpro.k001.danghuucanh.HelloWorld">
-        <constructor-arg value="inject     by    constructor" type="String"></constructor-arg>
-    </bean>
-
-</beans>
